requirements.txt
twikit
tweepy
python-dotenv
schedule
openai==0.28